Reworked blast processor recipe to be not crap - allows for easier implementation on the GUI
Removed old debug messages

Started work on GUI

+++ src/main/java/net/myriantics/klaxon/recipes/item_explosion_power/ItemExplosionPowerRecipe.java
@@ -64,8 +64,8 @@ public class ItemExplosionPowerRecipe implements Recipe<SimpleInventory> {
         return item;
     }
 
-    public float getExplosionPower() {
-        return (float) explosionPower;
+    public double getExplosionPower() {
+        return explosionPower;
     }
 
     public boolean producesFire() {
